started on sound design
