--- ARPG/Assets/DamageOverTime.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 35515e12e073e9b40a22cf7863eb9f1e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
