+++ ARPG/Assets/Maximo Animations/Boss Anim Controller.controller
@@ -353,7 +353,7 @@ AnimatorController:
     m_Type: 4
     m_DefaultFloat: 0
     m_DefaultInt: 0
-    m_DefaultBool: 1
+    m_DefaultBool: 0
     m_Controller: {fileID: 0}
   - m_Name: LightAttack
     m_Type: 4
