+++ ARPG/Assets/Scripts/NukeSound.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 46fb90ed62978994ebcb6a370ea6da0b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
